+++ client/src/Components/Buttons/index.js
@@ -1,0 +1,31 @@
+import React from 'react'
+
+const Buttons = () => {
+  return (
+    <>
+      <button type='button' className='btn'>
+        Basic
+      </button>
+      <button type='button' className='btn btn-default'>
+        Default
+      </button>
+      <button type='button' className='btn btn-primary'>
+        Primary
+      </button>
+      <button type='button' className='btn btn-success'>
+        Success
+      </button>
+      <button type='button' className='btn btn-info'>
+        Info
+      </button>
+      <button type='button' className='btn btn-warning'>
+        Warning
+      </button>
+      <button type='button' className='btn btn-danger'>
+        Danger
+      </button>
+    </>
+  )
+}
+
+export default Buttons
